+++ test/mix/tasks/volt_build_test.exs
@@ -20,6 +20,39 @@ defmodule Mix.Tasks.Volt.BuildTest do
     {:ok, tmp_dir: tmp_dir}
   end
 
+  test "--no-tree-shaking preserves unused exports", %{tmp_dir: tmp_dir} do
+    entry = Path.join(tmp_dir, "src/app.js")
+    outdir = Path.join(tmp_dir, "dist")
+
+    File.write!(Path.join(tmp_dir, "src/lib.js"), """
+    export function used() { return 'used' }
+    export function unused() { return 'unused' }
+    """)
+
+    File.write!(entry, """
+    import { used } from './lib.js'
+    console.log(used())
+    """)
+
+    Mix.Tasks.Volt.Build.run([
+      "--entry",
+      entry,
+      "--outdir",
+      outdir,
+      "--no-hash",
+      "--no-minify",
+      "--no-tailwind",
+      "--format",
+      "iife",
+      "--sourcemap",
+      "false",
+      "--no-tree-shaking"
+    ])
+
+    js_path = Path.join([outdir, "js", "app.js"])
+    assert File.read!(js_path) =~ "unused"
+  end
+
   test "--sourcemap false disables production sourcemaps", %{tmp_dir: tmp_dir} do
     entry = Path.join(tmp_dir, "src/app.js")
     outdir = Path.join(tmp_dir, "dist")